requirements.txt
Flask
groq
playwright
jieba
beautifulsoup4
python-dotenv
pydantic
google-generativeai
